(#37) Extract assertions to separate file
